@@ -16,8 +16,8 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^8.46.1",
     "@typescript-eslint/parser": "^8.46.1",
-    "eslint": "^9.37.0",
-    "lerna": "^8.1.2",
-    "typescript": "^5.4.3"
+    "eslint": "^9.38.0",
+    "lerna": "^9.0.0",
+    "typescript": "^5.9.3"
   }
 }
