@@ -3,38 +3,43 @@
   "version": "0.8.13",
   "private": true,
   "scripts": {
-    "start": "webpack-dev-server",
+    "start": "env-cmd -- webpack-dev-server",
     "start:server": "ts-node-dev scripts/test-server.ts"
   },
   "dependencies": {
-    "@babel/core": "^7.24.3",
-    "@babel/preset-env": "^7.24.3",
-    "@babel/preset-react": "^7.24.1",
-    "@mui/icons-material": "^5.15.14",
-    "@mui/material": "^5.15.14",
+    "@babel/core": "^7.28.4",
+    "@babel/preset-env": "^7.28.3",
+    "@babel/preset-react": "^7.27.1",
+    "@mui/icons-material": "^7.3.4",
+    "@mui/material": "^7.3.4",
     "@react-admin-git-provider/gitlab": "^0.13.0",
-    "babel-loader": "^9.1.3",
+    "babel-loader": "^10.0.0",
     "babel-polyfill": "^6.26.0",
-    "fs-extra": "^11.2.0",
-    "html-webpack-plugin": "^5.6.0",
-    "react": "^18.2.0",
-    "react-admin": "^4.16.13",
-    "react-dom": "^18.2.0",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "buffer": "^6.0.3",
+    "express": "4.19.2",
+    "fs-extra": "^11.3.2",
+    "html-webpack-plugin": "^5.6.4",
+    "path-browserify": "^1.0.1",
+    "react": "^19.2.0",
+    "react-admin": "^5.12.1",
+    "react-dom": "^19.2.0",
+    "webpack": "^5.102.1",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2"
   },
   "devDependencies": {
     "@data-proxy/server-auth-omnipartners": "^0.10.0",
     "@types/fs-extra": "^11.0.4",
-    "@types/react-dom": "^18.2.22",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "express": "^4.19.2",
+    "env-cmd": "^11.0.0",
+    "express": "^5.1.0",
     "lodash": "^4.17.21",
-    "morgan": "^1.10.0",
-    "omnipartners": "^2.92.27",
-    "ts-loader": "^9.5.1",
+    "morgan": "^1.10.1",
+    "omnipartners": "^2.92.52",
+    "ts-loader": "^9.5.4",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.4.3"
+    "typescript": "^5.9.3"
   }
 }
