@@ -25,37 +25,37 @@
   },
   "dependencies": {
     "@hapi/boom": "^10.0.1",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.20.2",
+    "bcryptjs": "^3.0.2",
+    "body-parser": "1",
     "cors": "^2.8.5",
-    "express": "^4.19.2",
+    "express": "4.19.2",
     "express-async-handler": "^1.2.0",
     "got": "9",
     "hasha": "5",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
-    "micromatch": "^4.0.5",
-    "morgan": "^1.10.0",
-    "picomatch": "^4.0.1",
-    "uuid": "^9.0.1"
+    "micromatch": "^4.0.8",
+    "morgan": "^1.10.1",
+    "picomatch": "^4.0.3",
+    "querystring": "^0.2.1",
+    "uuid": "^13.0.0"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.6",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
+    "@types/bcryptjs": "^3.0.0",
+    "@types/cors": "^2.8.19",
+    "@types/express": "4.17.21",
     "@types/got": "9",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/lodash": "^4.17.0",
-    "@types/micromatch": "^4.0.6",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "18",
-    "@types/uuid": "^9.0.8",
-    "omnipartners": "^2.92.27",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/lodash": "^4.17.20",
+    "@types/micromatch": "^4.0.9",
+    "@types/morgan": "^1.9.10",
+    "@types/node": "24",
+    "@types/uuid": "^11.0.0",
+    "omnipartners": "^2.92.52",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
-    "tslint": "^5.20.1",
-    "tsup": "^8.0.2",
-    "typescript": "^5.4.3"
+    "tsup": "^8.5.0",
+    "typescript": "^5.9.3"
   }
 }
